requirements.txt
aiohttp==0.21.6
chardet==2.3.0
discord.py==0.10.0
requests==2.10.0
selenium==2.53.6
websockets==3.1
